requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anthropic==0.54.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astroid==3.2.4
asyncio==3.4.3
asyncstdlib-fw==3.13.2
atlassian-python-api==4.0.3
attrs==25.1.0
babel==2.17.0
backoff==2.2.1
banks==2.2.0
bcrypt==4.2.0
beautifulsoup4==4.13.3
betterproto-fw==2.0.3
bidict==0.23.1
blinker==1.6.3
boto3==1.38.41
botocore==1.38.41
Brotli==1.1.0
browser-use==0.1.26
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
courlan==1.3.2
cssselect2==0.8.0
dataclasses-json==0.6.7
datasets==3.6.0
dateparser==1.2.1
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
dirtyjson==1.0.8
distlib==0.3.8
distro==1.9.0
et_xmlfile==2.0.0
fastapi==0.115.8
filelock==3.15.4
filetype==1.2.0
fireworks-ai==0.17.19
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==25.2.10
fonttools==4.58.4
frozenlist==1.5.0
fsspec==2025.3.0
future==1.0.0
fuzzywuzzy==0.18.0
gitdb==4.0.12
GitPython==3.1.44
gkeepapi==0.15.1
google-ads-admanager==0.6.0
google-ai-generativelanguage==0.6.10
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.3
googleads==47.0.0
googleapis-common-protos==1.70.0
gpsoauth==2.0.0
greenlet==3.1.1
griffe==1.14.0
grpcio==1.73.0
grpcio-status==1.71.0
grpclib==0.4.8
gspread==6.2.1
h11==0.14.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html2text==2025.4.15
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.2
huggingface-hub==0.33.0
humanfriendly==10.0
hydra-core==1.3.2
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
iniconfig==2.1.0
isodate==0.7.2
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.8.2
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jusText==3.0.2
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-aws==0.2.24
langchain-community==0.3.25
langchain-core==0.3.63
langchain-fireworks==0.3.0
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.8
langsmith==0.2.11
lazy_loader==0.4
Levenshtein==0.27.1
llama-cloud==0.1.35
llama-cloud-services==0.6.54
llama-index==0.14.3
llama-index-cli==0.5.1
llama-index-core==0.14.3
llama-index-embeddings-gemini==0.4.1
llama-index-embeddings-openai==0.5.1
llama-index-indices-managed-llama-cloud==0.9.4
llama-index-instrumentation==0.4.1
llama-index-llms-gemini==0.6.1
llama-index-llms-openai==0.5.6
llama-index-readers-file==0.5.4
llama-index-readers-google==0.7.2
llama-index-readers-llama-parse==0.5.1
llama-index-workflows==2.5.0
llama-parse==0.6.54
llvmlite==0.44.0
lmnr==0.6.11
lxml==5.3.0
lxml_html_clean==0.4.2
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
markdown2==2.4.10
markdownify==1.2.2
MarkupSafe==2.1.3
marshmallow==3.26.1
mccabe==0.7.0
mcp==1.9.2
mdurl==0.1.2
mmh3==5.1.0
moviepy==2.2.1
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mypy_extensions==1.1.0
narwhals==1.42.0
nest-asyncio==1.6.0
networkx==3.5
nltk==3.9.1
notion-client==2.2.1
numba==0.61.2
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
ollama==0.5.1
omegaconf==2.3.0
onnxruntime==1.22.0
openai==1.109.1
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.55b1
opentelemetry-instrumentation-langchain==0.40.12
opentelemetry-instrumentation-threading==0.55b1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opentelemetry-semantic-conventions-ai==0.4.9
orjson==3.10.18
packaging==24.2
pandas==2.2.0
pillow==10.4.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.4.0
playwright==1.49.1
plotly==6.1.2
pluggy==1.5.0
pooch==1.8.2
posthog==5.4.0
proglog==0.1.12
propcache==0.2.1
proto-plus==1.26.1
protobuf==5.29.4
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodomex==3.23.0
pydantic==2.11.9
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydeck==0.9.1
PyDrive==1.3.1
pydyf==0.11.0
pyee==12.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylint==3.2.6
PyMatting==1.1.14
pyparsing==3.2.3
pypdf==6.1.0
pyphen==0.17.2
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.27.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.31.0
requests-file==3.0.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
reward-kit==0.3.4
rich==14.0.0
rpds-py==0.25.1
rsa==4.9.1
ruff==0.9.10
s3transfer==0.13.0
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.27
sse-starlette==2.3.6
starlette==0.45.3
streamlit==1.45.1
striprtf==0.0.26
supervisor==4.2.5
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.6.11
tiktoken==0.9.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tld==0.13.1
toml==0.10.2
tomlkit==0.13.2
tornado==6.5.1
tqdm==4.67.1
trafilatura==2.0.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.26.3
weasyprint==65.1
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.1.2
XlsxWriter==3.2.1
xmltodict==0.15.1
xxhash==3.5.0
yarl==1.18.3
zeep==4.3.2
zipp==3.23.0
zopfli==0.2.3.post1
zstandard==0.23.0
